# src/test/resources/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "HRD Room Service API",
    "description" : "REST API for HRD room service operations, including authentication, academic years, classrooms, courses, and enrollments. Obtain a JWT from POST /api/v4/auth/login; send it as Bearer authentication on protected routes.",
    "contact" : {
      "name" : "HRD Room Service Team",
      "email" : "penhseyha4980@gmail.com"
    }
  },
  "paths" : {
    "/api/v4/enrollments/{enrollmentId}/terminate" : {
      "put" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "terminate",
        "parameters" : [ {
          "name" : "enrollmentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TerminateEnrollmentRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/enrollments/{enrollmentId}/reactivate" : {
      "put" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "reactivate",
        "parameters" : [ {
          "name" : "enrollmentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/enrollments" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "listAll",
        "parameters" : [ {
          "name" : "includeArchived",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "enroll",
        "parameters" : [ {
          "name" : "studentId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "courseId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/enrollments/move-to-advanced" : {
      "post" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "moveToAdvanced",
        "parameters" : [ {
          "name" : "studentId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "basicCourseId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "advancedCourseId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/enrollments/{enrollmentId}" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "byId_2",
        "parameters" : [ {
          "name" : "enrollmentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/enrollments/student/{studentId}" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "byStudent",
        "parameters" : [ {
          "name" : "studentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "includeArchived",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/enrollments/student/{studentId}/by-course-type" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "byStudentGrouped",
        "parameters" : [ {
          "name" : "studentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "includeArchived",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentByCourseTypeResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/enrollments/course/{courseId}" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "byCourse",
        "parameters" : [ {
          "name" : "courseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/enrollments/by-course-type" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "byCourseType",
        "parameters" : [ {
          "name" : "includeArchived",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEnrollmentByCourseTypeResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/classrooms/{classroomId}/students" : {
      "get" : {
        "tags" : [ "classroom-controller" ],
        "operationId" : "students",
        "parameters" : [ {
          "name" : "classroomId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListUUID"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "TerminateEnrollmentRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseEnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/EnrollmentResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "EnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "enrollmentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "studentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "courseId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "courseName" : {
            "type" : "string"
          },
          "courseType" : {
            "type" : "string"
          },
          "academicYearId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "academicYearName" : {
            "type" : "string"
          },
          "generation" : {
            "type" : "integer",
            "format" : "int32"
          },
          "enrolledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isPassed" : {
            "type" : "boolean"
          },
          "isTerminated" : {
            "type" : "boolean"
          },
          "terminatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "terminationReason" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ApiResponsePageResponseEnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseEnrollmentResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "PageResponseEnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EnrollmentResponse"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponseEnrollmentByCourseTypeResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/EnrollmentByCourseTypeResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "EnrollmentByCourseTypeResponse" : {
        "type" : "object",
        "properties" : {
          "basic" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EnrollmentResponse"
            }
          },
          "advanced" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EnrollmentResponse"
            }
          }
        }
      },
...
      "ApiResponseListUUID" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}